+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPQRMTMV0JJ55WRXVVDFA2ES"
+        "01KPQZ9EWE915VFG7476QVRAA5"
     )
     r"""The unique identifier of the data source"""