+++ packages/orq-rc/RELEASES.md
@@ -7149,3 +7149,13 @@ Based on:
 - [python v4.8.0-rc.37] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.37] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.37 - packages/orq-rc
+
+## 2026-04-21 12:14:58
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.38] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.38] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.38 - packages/orq-rc